combine User Resources to one java class
