+++ src/main/java/com/ineuron/INeuronApplication.java
@@ -1,7 +1,7 @@
 package com.ineuron;
 
 import com.ineuron.health.TemplateHealthCheck;
-import com.ineuron.resources.user.*;
+import com.ineuron.resources.*;
 import io.dropwizard.Application;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
@@ -30,23 +30,19 @@ public class INeuronApplication extends Application<INeuronConfiguration> {
         @Override
         public void run(INeuronConfiguration configuration,
                         Environment environment) {
-        	final AuthenticateResource resource = new AuthenticateResource(
+        	final UserResource userResource = new UserResource(
         	        configuration.getText(),
         	        configuration.getDefaultName()
         	    );
-        	
-        	final RegisterResource regResource = new RegisterResource(
-        	        configuration.getText(),
-        	        configuration.getDefaultName()
-        	    );
+        
         	
     	final TemplateHealthCheck healthCheck =
     	        new TemplateHealthCheck(configuration.getText());
     	
     	    environment.healthChecks().register("text", healthCheck);
 
-    	    environment.jersey().register(resource);
-    	    environment.jersey().register(regResource);
+    	    environment.jersey().register(userResource);
+    	 
     }
 
 }
